requirements_for_render.txt
discord.py>=2.6.0
aiosqlite>=0.19.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0